Correcting a typo in the argument.
Earlier today when submitting this assignment, I accidentally submitted
a version of this script with a typo in the argument. I was cleaning up
the script to make it more readable and changed the argument to
“output” whereas previously I was using “x,” and I forgot to make the
change in all locations. This commit updates “x” to “output” in three
additional occurrences. You’ll see that this version is otherwise
identical to the one submitted earlier today.